fix: DS2-H1 exposure all-time scan; DS2-M1 pair-level drift; DS2-M2 weighted entry_price; DS2-L1/L2/L3 audit/pnl/summary; DA-M9/M10/M11/L9/L10/L11/L12 data analyst fixes

+++ agents/ruppert/data_analyst/market_cache.py
@@ -141,8 +141,8 @@ def get_market_price(ticker: str, fallback_client=None) -> dict | None:
                     return {
                         'yes_bid': bid,
                         'yes_ask': ask,
-                        'no_bid':  round((1 - rest_ask_d) * 100),
-                        'no_ask':  round((1 - rest_bid_d) * 100),
+                        'no_bid':  int(round((1 - rest_ask_d) * 100)),
+                        'no_ask':  int(round((1 - rest_bid_d) * 100)),
                         'source': 'rest',
                     }
                 # REST returned market but null prices — fall through to stale
